requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
Django==3.2
django-allauth==0.51.0
django-filter==22.1
django-phonenumber-field==7.0.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
flake8==6.0.0
idna==3.4
importlib-metadata==5.1.0
isort==5.10.1
Markdown==3.4.1
mccabe==0.7.0
oauthlib==3.2.2
phonenumberslite==8.13.1
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
zipp==3.11.0
